% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anyBaseToDecimal}
\alias{anyBaseToDecimal}
\title{Convert from anybase to decimal}
\usage{
anyBaseToDecimal(value, symbols = LETTERS, zero_indexed = F)
}
\arguments{
\item{value}{The starting value}

\item{symbols}{A vector containing all of the symbols used in the origin place
value system listed in sequential order.}

\item{zero_indexed}{Logical. Set to TRUE if the first value in \code{symbols}
represents 1, FALSE if 0.}
}
\value{
\code{value} converted to decimal
}
\description{
Convert from anybase to decimal
}
\examples{
anyBaseToDecimal("A")
anyBaseToDecimal("ABC")
anyBaseToDecimal("FF", symbols = c(0:9, LETTERS[1:6]), zero_indexed=T) #hex to decimal
}
