% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaPhotoBoothShine.R
\name{coaPhotoBoothShine}
\alias{coaPhotoBoothShine}
\title{CoA Bar Plot for Shiny}
\usage{
coaPhotoBoothShine(tib, y_var = "conc_incell_uM", grouping_var = "tx_run",
  analytes = c("Acetyl-CoA", "Isobutyryl-CoA", "Propionyl-CoA"),
  species = "both", limits = list(Acetyl = c(0, 55), Isobutyryl = c(0, 45),
  Propionyl = c(0, 65)), x_max = Inf)
}
\arguments{
\item{tib}{A tibble or data.frame with PureHoney data including vars(tx_cmpd, curve_plot).}

\item{y_var}{A character with the column name to be used for the y-axis.}

\item{grouping_var}{A character string specifying the column in \code{tib} that specifies the experimental grouping.}

\item{analytes}{A character vector providing the analytes to be plotted.}

\item{species}{A character describing which labeled analytes to plot. Accepts "both" (default), "C13", or "C12".}

\item{limits}{A named list with the names matching \code{unique(tib$curve_plot)} and values of numeric vectors with length of 2, describind the y-axis bound for each `curve_plot`.}

\item{x_max}{Numeric. The highest dose considered in the calculation.}
}
\value{
A cowplot with analytes as rows and treatments as columns
}
\description{
Plot each cmpd from a PureHoney tibble for the hemoshine server.
}
\examples{
pah <- filter(samps2, run == "PAH0503") # tib
coaPhotoBooth(pah)
}
