% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaPhotoBoothShiny.R
\name{coaPhotoBoothShiny}
\alias{coaPhotoBoothShiny}
\title{CoA Bar Plot for Shiny}
\usage{
coaPhotoBoothShiny(tib, y_var = "conc_incell_uM", grouping_var = "tx_cmpd",
  species = "both", limits = list(Acetyl = c(0, 55), Isobutyryl = c(0, 45),
  Propionyl = c(0, 65)))
}
\arguments{
\item{tib}{A tibble or data.frame with PureHoney data including vars(tx_cmpd, curve_plot).}

\item{y_var}{A character with the column name to be used for the y-axis.}

\item{species}{A character describing which type carbon species to plot. Accepts "both" (default), "heavy", or "natural".}

\item{limits}{A named list with the names matching \code{unique(tib$curve_plot)} and values of numeric vectors with length of 2, describind the y-axis bound for each `curve_plot`.}
}
\value{
A cowplot with analytes as rows and treatments as columns
}
\description{
Plot each cmpd from a PureHoney tibble separatly in a (n-curve)x1 \code{cowplot::plot_grid()} with the standard RNO theme.
}
\examples{
pah <- filter(samps2, run == "PAH0503") # tib
coaPhotoBooth(pah)
}
