% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drcLabels}
\alias{drcLabels}
\title{Overwrite lowest value label with "0" for dose response curves}
\usage{
drcLabels(breaks)
}
\arguments{
\item{limits}{Default \code{ggplot2} numeric object of length 2, defining upper and lower bounds for a scale.}
}
\value{
A numeric vector with the the lowest value set to 0.
}
\description{
A helper function to be used in \code{ggplot() + scale_x_continuous(..., labels = drcLabels)} for relabeling the lowest compound concentration from its log-approximate value. This function is the correction for \code{newZeros()}.
}
\examples{
ggplot(data = ph_drc, aes(tx_conc, conc_incell_uM) +
geom_curve() +
scale_x_continuous(breaks = drcBreaks, labels = drcLabels)
}
