% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcPhotoBoothShine.R
\name{drcPhotoBoothShine}
\alias{drcPhotoBoothShine}
\title{CoA DRC Curve Plots}
\usage{
drcPhotoBoothShine(tib, analytes = c("Acetyl-CoA", "Isobutyryl-CoA",
  "Propionyl-CoA"), y_var = "conc_incell_uM", limits = NULL, x_max = Inf,
  grouping_var = "tx_run", drm_error = FALSE, ec50 = TRUE,
  ec50ci = FALSE, Hill = FALSE, up_asym = FALSE, low_asym = FALSE,
  robust = TRUE)
}
\arguments{
\item{tib}{A tibble or data.frame with PureHoney data including vars(tx_cmpd, curve_plot).}

\item{analytes}{A character vector with a subset of \code{unique(tib$targ)} for the target/analytes to be plotted.}

\item{y_var}{A character with the \code{tib} column name to be used for the y-axis. Default is "conc_incell_uM", "conc_corrected" may also be useful.}

\item{limits}{A named list with the names matching \code{unique(tib$curve_plot)} and values of numeric vectors with length of 2, describind the y-axis bound for each `curve_plot`. If there is a target present in \code{targs} that is not in \code{limits} the limits will be calculated with \code{ggplot2}'s defualt behavior.}

\item{x_max}{Numeric. The highest dose considered in the calculation.}

\item{grouping_var}{A character string representing the column used for grouping experiments.}

\item{drm_error}{Logical. Determines if drc() convergence failure results in a error (TRUE) or warning (FALSE)}

\item{ec50}{Logical. Show EC50.}

\item{ec50ci}{Logical. Show 95\% CI of EC50.}

\item{Hill}{Logical. Show Hill coefficient.}

\item{up_asym}{Logical. Show the upper asymptote.}

\item{low_asym}{Logical. Show the lower asymptote.}

\item{robust}{Logical. Use robust curve estimation.}

\item{deriv}{Logical. Use derivative in DRC estimation.}
}
\value{
A plot object.
}
\description{
Plot one or more dose response curves.
}
\examples{
pah <- filter(samps2, run == "PAH0503") # tib
drcPhotoBooth(pah)
}
