% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drcs}
\alias{drcs}
\title{Fitted dose response models}
\format{A list with three \code{drc} objects}
\usage{
drcs
}
\description{
Three model fits for Isobutryl-CoA dose response from PureHoney plate 547.
}
\keyword{datasets}
