% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcTableShine.R
\name{drcTableShine}
\alias{drcTableShine}
\title{CoA DRC Tables}
\usage{
drcTableShine(tib, analytes = c("Acetyl-CoA", "Isobutyryl-CoA",
  "Propionyl-CoA"), x_max = Inf, y_var = "conc_incell_uM",
  grouping_var = "tx_run", drm_error = FALSE, ec50 = TRUE, Hill = FALSE,
  up_asym = FALSE, low_asym = FALSE, cis = FALSE, robust = TRUE)
}
\arguments{
\item{tib}{A tibble or data.frame with PureHoney data including vars(tx_cmpd, curve_plot).}

\item{analytes}{A character vector of the target/analytes to be plotted.}

\item{x_max}{Numeric. The highest dose considered in the calculation.}

\item{y_var}{A character with the \code{tib} column name to be used for the y-axis. Default is "conc_incell_uM", "conc_corrected" may also be useful.}

\item{grouping_var}{A character string representing the column used for grouping experiments.}

\item{drm_error}{Logical. Determines if drc() convergence failure results in a error (TRUE) or warning (FALSE)}

\item{ec50}{Logical. Show EC50.}

\item{Hill}{Logical. Show Hill coefficient.}

\item{up_asym}{Logical. Show the upper asymptote.}

\item{low_asym}{Logical. Show the lower asymptote.}

\item{cis}{Logical. Show 95\% CI of stats.}

\item{robust}{Logical. Use robust curve estimation.}
}
\value{
A data frame.
}
\description{
Generate a DRC stats table.
}
\examples{
pah <- filter(samps2, run == "PAH0503") # tib
drcPhotoBooth(pah)
}
