% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fillNAs}
\alias{fillNAs}
\title{Fill NAs in vector with neighboring values}
\usage{
fillNAs(vector, reverse = F)
}
\arguments{
\item{vector}{The starting vector}

\item{reverse}{Logical. The direction to traverse \code{vector}}
}
\value{
The back (forward) filled vector.
}
\description{
Back- or forward-fill NAs in a vector
}
\examples{
x <- c("A", rep(NA, 8), "Z")
fillNAs(x)
fillNAs(x, reverse=T)
}
