% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric mean}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{A numeric vector whose geometric mean is to be calulated}
}
\value{
The geometric mean of \code{x}.
}
\description{
Calculate the geometric mean.
}
\examples{
x <- rpois(10, lambda = 10)
geom_mean(x)
}
