% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_mean_se}
\alias{geom_mean_se}
\title{Geometric standard error of the mean}
\usage{
geom_mean_se(x, mult = 1)
}
\arguments{
\item{x}{A numeric vector whose geometric standard error of the mean is to be calulated.}
}
\value{
The geometric standard error of the mean of \code{x}.
}
\description{
Calculate the geometric standard error of the mean.
}
\examples{
x <- rpois(10, lambda = 10)
geom_mean_se(x)
}
