% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{getCurves}
\alias{getCurves}
\title{Generate y values in increments of a given x range.}
\usage{
getCurves(range, fit, logrange = T, npoint = 100)
}
\arguments{
\item{range}{A 2-element vector giving the x range.}

\item{fit}{An object of class drc fit with \code{drm()}.}

\item{logrange}{Logical. If TRUE increments are in log space.}

\item{npoint}{Number of increments in the x range.}
}
\value{
A data frame of x and y values.
}
\description{
Generate y values in increments of a given x range.
}
\examples{
fit <- drc::drm(disp~wt, data=mtcars, fct=LL.4())
curve <- getCurves(c(2,5), fit, logrange = F)
plot(ys~xs, data=curve)
}
