% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{getEC50}
\alias{getEC50}
\title{Get the EC50 value of a 4-parmeter log-logistic function}
\usage{
getEC50(fit, CI95 = FALSE)
}
\arguments{
\item{fit}{An object of class drc fit with \code{drm()} and the argument \code{fct = LL.4()}.}

\item{CI95}{Logical. Report 95\% confidence interval.}
}
\value{
The x value corresponding to point at which the function is rotationally symmetric.
}
\description{
Get the EC50 value of a 4-parmeter log-logistic function
}
\examples{
fit <- drc::drm(disp~wt, data=mtcars, fct=drc::LL.4())
getEC50(fit)
}
