% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{getHillSlope}
\alias{getHillSlope}
\title{Get the Hill coefficient of a 4-parmeter log-logistic function}
\usage{
getHillSlope(fit, curve_id = NULL)
}
\arguments{
\item{fit}{An object of class drc fit with \code{drm()} and the argument \code{fct = LL.4()}.}

\item{curve_id}{A character string identifying the curve if multiple curves were fit by \code{drm()}.}
}
\value{
The Hill coefficient:
\eqn{y=c+\frac{d-c}{1+10^{(ln(\tilde{e})-x) \times Hill}}}.
Where c = upper asymptote, d = lower asymptote, and \eqn{\tilde{e}} = the EC50 value.
}
\description{
Get the Hill coefficient of a 4-parmeter log-logistic function
}
\examples{
fit <- drc::drm(disp~wt, data=mtcars, fct=drc::LL.4())
getEC50(fit)
}
