% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{getLowerAsym}
\alias{getLowerAsym}
\title{Get the y value of the lower asymptote in a 4-parmeter log-logistic function}
\usage{
getLowerAsym(fit, CI95 = FALSE)
}
\arguments{
\item{fit}{An object of class drc fit with \code{drm()} and the argument \code{fct = LL.4()}.}

\item{CI95}{Logical. Report 95\% confidence interval.}
}
\value{
The y intercept of the lower asymptote.
}
\description{
Get the y value of the lower asymptote in a 4-parmeter log-logistic function
}
\examples{
fit <- drc::drm(disp~wt, data=mtcars, fct=drc::LL.4())
getLowerAsym(fit)
}
