% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaData}
\alias{getMetaData}
\title{Import meta data}
\usage{
getMetaData(study_id, fully_parse = TRUE, static = TRUE)
}
\arguments{
\item{study_id}{Character vector in the standard study name format (i.e. "RNO0101-1" or "RNO0101").}

\item{fully_parse}{Logical indicating whether to try and filter the database to a specific run; Set \code{FALSE} to troubleshoot.}
}
\value{
A data frame containing the meta information from a given study or a large data frame with all the meta information from all studies in a program.
}
\description{
Reads meta data from sample databases on Box via \code{boxr::box_read_excel()}
}
\examples{
# returns only meta data from study RNO0738
meta <- getMetaData("RNO0738")
meta <- getMetaData(c("HEM0100-5", "HEM0100-6"))

# returns all studies' meta data from RNO07 project
all_meta <- getMetaData("RNO0738", fully_parse = F)
}
