% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{getYBounds}
\alias{getYBounds}
\title{Get the bottom and top of a log-logistic curve in a given x range.}
\usage{
getYBounds(fit, high_x, low_x)
}
\arguments{
\item{fit}{An object of class drc fit with \code{drm()}.}

\item{high_x}{Numeric. The high end of the x range.}

\item{low_x}{Numeric. The low end of the x range.}
}
\value{
A named vector reporting the top and bottom of the curve (toc, boc)
}
\description{
Get the bottom and top of a log-logistic curve in a given x range.
}
\examples{
fit <- drc::drm(disp~wt, data=mtcars, fct=drc::LL.4())
getYBounds(fit, 2, 5)
}
