% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{inversePredLL4}
\alias{inversePredLL4}
\title{Predict x from y in a 4-parameter log-logistic model}
\usage{
inversePredLL4(y, fit)
}
\arguments{
\item{y}{The y value.}

\item{fit}{An object of class drc fit with \code{drm()} and the argument \code{fct = LL.4()}.}
}
\value{
The x value corresponding to the given y.
}
\description{
Predict x from y in a 4-parameter log-logistic model
}
\examples{
fit <- drc::drm(mpg~hp, data=mtcars, fct=LL.4())
inversePredLL4(15, fit)
}
