% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{inversePredLm}
\alias{inversePredLm}
\title{Predict x from y in a linear model}
\usage{
inversePredLm(y, fit)
}
\arguments{
\item{y}{The y value.}

\item{fit}{An object of class lm with an intercept and a single coefficient.}
}
\value{
The x value corresponding to the given y.
}
\description{
Predict x from y in a linear model
}
\examples{
fit <- lm(mpg~hp, data=mtcars)
inversePredLm(15, fit)
}
