% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCIs.R
\name{makeCIs}
\alias{makeCIs}
\title{Build plot-ready confidence intervals}
\usage{
makeCIs(mod, level = 0.95, all = FALSE)
}
\arguments{
\item{mod}{A model object of class "lm", "lmerMod", or "lmerModLmerTest".}

\item{level}{Numeric (0-1). The confidence level required.}

\item{all}{Logical. Return only the first fixed effect on RHS of the model formula (FALSE) or all fixed effects (TRUE).}
}
\description{
Wrapper for \code{confint()} that handles the output munging for models of class "lm", "lmerMod" and "lmerModLmerTest".
}
\examples{
mod <- lm(Sepal.Length ~ Species, data = iris)
ci <- makeCIs(mod)

mods <- mtcars \%>\% split(.$cyl) \%>\% purrr::map(~ lm(mpg ~ hp, data = .))
cis <- purrr::map(mods, makeCIs)
}
