% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLogFCs.R
\name{makeLogFCs}
\alias{makeLogFCs}
\title{Calculate log fold changes and associated p-values.}
\usage{
makeLogFCs(ht, base = 10, round_to = 2, conf_level = 0.95)
}
\arguments{
\item{ht}{An object of class "glht" from \code{multcomp::glht}.}

\item{base}{A numeric value specifying the log based used in the model. Default is 10.}

\item{round_to}{A numeric value specifying the number of decimal places in percent change columns in result.}

\item{conf_level}{A numeric value between 0 and 1 for confidence level. Default is .95.}
}
\value{
A data frame with est (estimate change), lwr, upr (confidence bounds), contrast (predictor levels being compared), pct_label (text, for plotting labels), pct_fill (numeric, for plotting scales), and a p-value adjusted for multiple comparrisons but \code{multcomp::glht}.
}
\description{
Wrapper for calculating fold changes based on \code{multcomp::glht()} object.
}
\examples{
mod <- lm(Sepal.Length ~ 0 + Species, data = iris) # use model with intercept set to 0
ht <- multcomp::glht(mod, linfct = mcp(Species = "Dunnet"))
fcs <- makeLogFCs(ht)
}
