% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeSerialDilution}
\alias{makeSerialDilution}
\title{Make a vector of serially halved values.}
\usage{
makeSerialDilution(highest, number)
}
\arguments{
\item{highest}{The highest value in the geometric progression.}

\item{number}{The number of values in addition to \code{highest} in the sequence.}
}
\value{
A vector containing the geometric progression
}
\description{
Make a vector of serially halved values.
}
\examples{
makeSerialDilution(10, 5)
}
