% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{meltPlate}
\alias{meltPlate}
\title{Melt a data frame from plate format to long format}
\usage{
meltPlate(plt)
}
\arguments{
\item{plt}{A data frame representing an assay plate.}
}
\value{
The melted data frame.
}
\description{
Melt a data frame from plate format to long format
}
\examples{
plate <- as.data.frame(matrix(rnorm(96), nrow = 8))
meltPlate(plate)
}
