% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{meltPlateXlsx}
\alias{meltPlateXlsx}
\title{Read xlsx data in plate format and convert to long format}
\usage{
meltPlateXlsx(f, sheet, rows, cols)
}
\arguments{
\item{f}{A character string -- the xlsx file.}

\item{sheet}{The name or index of the sheet.}

\item{rows}{a numeric vector of the rows to read}

\item{cols}{a numeric vector of the columns to read}
}
\value{
A melted data frame.
}
\description{
Read in plate format data from an xlsx file and return a melted
data frame with the contents and well coordinates
}
