% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newZeros.R
\name{newZeros}
\alias{newZeros}
\title{Convert concentrations of zero to non-zero values in dose response curves}
\usage{
newZeros(vec, is_log = F)
}
\arguments{
\item{vec}{A numeric vector with the concentrations in a dose response curve}

\item{is_log}{Default is FALSE. Are the concentration values in `vec` log transformed?}
}
\value{
A numeric value to be assigned as the new non-zero concentration to represent actual concentrations of zero.
}
\description{
Tool to adjust concentration values of zero to non-zero values based on the average gap between concentraions in dose response curve. Adjust zero concentrations to be two average gaps less than the lower non-zero concentration. Useful for plotting on with `scale_x_log10()`
}
\examples{
drc_concs <- c(0, .3, 1, 3, 10, 30, 100)
drc_concs[drc_concs == 0] <- newZeros(drc_concs)
drc_concs
}
