% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normToAcetyl.R
\name{normToAcetyl}
\alias{normToAcetyl}
\title{Normalize to acetyl-CoA}
\usage{
normToAcetyl(x)
}
\arguments{
\item{x}{A tibble or data frame with PureHoney data including vars(tx_cmpd, curve_plot).}
}
\value{
A data frame.
}
\description{
Create a column in a PH analysis, \code{to_acoa_ratio}, that gives the ratio of analytes to acetyl-CoA
}
