% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outputPlotsAsPdfs}
\alias{outputPlotsAsPdfs}
\title{Export plots}
\usage{
outputPlotsAsPdfs(plot_list, pdf_path, overwrite = TRUE,
  session_info = TRUE, height = 8.5, width = 11)
}
\arguments{
\item{plot_list}{List containing plot objects.}

\item{pdf_path}{Path for output pdf that will be created. Should end in ".pdf".}

\item{overwrite}{Logical indicating whether to overwrite an existing directory with the same path. Defaults to \code{TRUE} for easy analysis interation. Adjust to \code{FALSE} to add individual images to existing directory, may throw errors about existing files.}

\item{session_info}{Logical whether to include a \code{gtools::textplot} with the output of \code{sessionInfo()}. Useful for reproducible research.}

\item{height}{Numeric object describing height of graphics device. Default unit is inches.}

\item{width}{Numeric object describing width of graphics device. Default unit is inches.}
}
\value{
A directory with the elements of \code{plot_list} as png files
}
\description{
Takes a list of plot objects and builds a pdf document with those plots.
}
\examples{
plots <- mtcars \%>\% split(.$cyl) \%>\%
           purrr::map(~ ggplot(., aes(mpg, disp)) +
                          geom_point())
outputPlotsAsPngs(plots, "~/Desktop/mtcars_plots_by_cyl/")
}
