% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outputPlotsAsPngs}
\alias{outputPlotsAsPngs}
\title{Export plots}
\usage{
outputPlotsAsPngs(plot_list, dir_path, overwrite = TRUE,
  session_info = TRUE)
}
\arguments{
\item{plot_list}{List containing plot objects}

\item{dir_path}{Path for output directory that will be created}

\item{overwrite}{Logical indicating whether to overwrite an existing directory with the same path. Defaults to \code{TRUE} for easy analysis interation. Adjust to \code{FALSE} to add individual images to existing directory, may throw errors about existing files.}

\item{session_info}{Logical whether to include a \code{gtools::textplot} with the output of \code{sessionInfo()}. Useful for reproducible research.}
}
\value{
A directory with the elements of \code{plot_list} as png files
}
\description{
Takes a list of plot objects and builds a folder with those plots as individual png files
}
\examples{
plots <- mtcars \%>\% split(.$cyl) \%>\%
           purrr::map(~ ggplot(., aes(mpg, disp)) +
                          geom_point())
outputPlotsAsPngs(plots, "~/Desktop/mtcars_plots_by_cyl/")
}
