% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{predDRC}
\alias{predDRC}
\title{Predict y from x in a log-logistic model}
\usage{
predDRC(x, fit)
}
\arguments{
\item{x}{The x value.}

\item{fit}{An object of class drc fit with \code{drm()}.}
}
\value{
The y value corresponding to the given x.
}
\description{
Predict y from x in a log-logistic model
}
\examples{
fit <- drc::drm(mpg~hp, data=mtcars, fct=LL.4())
predDRC(150, fit)
}
