% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCFX.R
\name{readCFX}
\alias{readCFX}
\title{Generic parser for Biorad CFX thermocycler}
\usage{
readCFX(file, capture = "")
}
\arguments{
\item{file}{Path to a BioRad CFX output csv.}

\item{capture}{A single capture regex for extracting plate id from file.}
}
\description{
Reads the defualt output csv from CFX Manager 3.1.1621.0826
}
\examples{
f <- boxr::box_dl(198532377451, local_dir = "~/.Trash/") \%>\%
       readCFX(capture = ".*_(\\\\d)\\\\.csv")

}
