% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFLUOstar.R
\name{readFLUOstar}
\alias{readFLUOstar}
\title{Generic parser for Omega FluoStar plate analyzer}
\usage{
readFLUOstar(file, capture = "", col_drops = NA)
}
\arguments{
\item{file}{Path to a BMG FLUOstart Omega plate reader output xlsx.}

\item{capture}{A single capture regex for extracting plate id from file.}

\item{col_drops}{An optional character vector with the names of columns to be dropped.}
}
\description{
Versital parser that orders standard columns (ie raw, blank_raw, machine_fit, etc)
automatically for \code{rbind()}-ing. Verbose with warnings.
}
\examples{
f <- boxr::box_dl(198532377451, local_dir = "~/.Trash/") \%>\%
       readFLUOstar(capture = ".*_(\\\\d)\\\\.csv")

}
