% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPHRaw.R
\name{readPHRaw}
\alias{readPHRaw}
\title{PureHoney Assay Result Reader}
\usage{
readPHRaw(raw_file, plate_id)
}
\arguments{
\item{raw_file}{Path to .csv file with PureHoney raw output.}

\item{plate_id}{Character vector describing the plate id. Typically just the full file name.}
}
\description{
High level wrapper function that that imports the raw data output from PureHoney as a tibble.
}
\examples{
raw <- dir(".", pattern = "\\\\.csv", full.names = T) \%>\%
         map_df(~ readPHRaw(., plate_id = .))
}
