% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{robusty}
\alias{robusty}
\title{Bastard child of \code{robustifyDRC()}}
\usage{
robusty(fit, conv = 0.01, maxits = 100, verbose = TRUE)
}
\arguments{
\item{conv}{Numeric value representing the convergence criterion (manhattan distance between weight vectors)}

\item{maxits}{Numeric value representing maximum number of iterations}

\item{verbose}{Logical. If TRUE reports the manhattan distance for each iteration.}
}
\description{
While loop formulation of \code{robustifyDRC()} to eleminate scoping woes and pasing of formula in arguments
}
