% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_assayr.R
\name{theme_assayr}
\alias{theme_assayr}
\title{Complete theme for assay report default}
\usage{
theme_assayr(base_size = 18, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
This is a complete theme build off of \code{ggplot2::theme_classic()}. See \code{?theme_classic} for details.
}
\examples{
ggplot(mtcars, aes(mpg, hp, color = as.factor(cyl))) +
geom_point() +
theme_assayr()
}
