% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMolDev.R
\name{readMolDev}
\alias{readMolDev}
\title{Molecular Devices Scope Output Reader}
\usage{
readMolDev(scope_txt, measurements = c("stell  nuc count", "mac nuc count",
  "smaa area", "integrated int", "Nuclear Count", "Vesicle", "Cell Count"))
}
\arguments{
\item{scope_txt}{A path to a .txt or .xlsx file with Molecular Devices scope output}

\item{measurements}{A character vector with the column names of measurments to extract. By default this is set to include all common measurements. Will drop measurements if they are not present.}
}
\value{
A tidy data frame with columns for plate, well, site and measurments
}
\description{
High level wrapper that imports a Molecular Devices scope output file as a data frame. Uses specific document markers to extract relevant meta information, vunerable to changes in the scope output template.
}
\examples{
readMolDev("BGA0101-1 SMAA.txt")
readMolDev("BGA0101-1 NileRed Nuc Count.xlsx")
}
