% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMolDev.R
\name{readMolDev}
\alias{readMolDev}
\title{Parser for Molecular Devices Output}
\usage{
readMolDev(scope_txt, measurements = c("stell  nuc count", "mac nuc count",
  "smaa area", "integrated int", "Nuclear Count", "Vesicle", "Cell Count"))
}
\arguments{
\item{scope_txt}{A valid path to a .txt or .xlsx file with Molecular Devices scope output.}

\item{measurements}{A character vector with the column names of measurments to extract. Default is set to include all common measurements. Drops measurements if not available in scope_txt.}
}
\value{
A tibble with columns for plate, well, site, uid and all measurments available.
}
\description{
Convenience wrapper for importing a Molecular Devices scope output file as a data frame.
}
\examples{
readMolDev("BGA0101-1 SMAA.txt")
readMolDev("BGA0101-1 NileRed Nuc Count.xlsx")
}
