% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPHAssay.R
\name{readPHAssay}
\alias{readPHAssay}
\title{PureHoney Assay Layout Reader}
\usage{
readPHAssay(layout_file, plate_ids = NULL, meta_cols = c(1, 3:5),
  meta_names = c("contents", "sample_id", "imaging_plate", "imaging_well"))
}
\arguments{
\item{layout_file}{Path to .xlsx file with PureHoney layout and sample data.}

\item{plate_ids}{Optional argument to extract only specific plate(s) from multi-plate document. Expects character vector.}

\item{meta_cols}{Numeric vector defining the columns to extract for meta section. Default is most common arrangment.}

\item{meta_names}{Character vector defining column names for meta section. Length must match \code{legnth(meta_cols)}. Default is most common arrangment.}
}
\description{
High level wrapper function that that imports the assay layout for PureHoney as a tibble. Uses fixed offset from document markers to capture specific regions, prone to errors if assay document template is altered.
}
\examples{
meta <- readPHAssay("CoA plates 100000537_100000538_RNO02004_PAH0119_6-29-17.xlsx")
}
