% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{robustifyDrc}
\alias{robustifyDrc}
\title{Make a robust DRC estimate based on Nguyen et al. 2014}
\usage{
robustifyDrc(fit, conv = 0.01)
}
\arguments{
\item{fit}{An object of class drc fit with \code{drm()} and the argument \code{fct = LL.4()}.}

\item{conv}{Numeric value representing the convergence criterion (manhattan distance between weight vectors)}
}
\value{
An object of class drc
}
\description{
Make a robust DRC estimate based on Nguyen et al. 2014
}
\examples{
fit <- drc::drm(disp~wt, data=mtcars, fct=LL.4())
robustifyDrc(fit)
}
